+++ src/registry/test_utils.rs
@@ -154,6 +154,7 @@ pub fn create_test_registry_with(
         .enable_blob_redirect(global.resolved_enable_blob_redirect())
         .enable_manifest_redirect(global.resolved_enable_manifest_redirect())
         .max_manifest_size_bytes(global.max_manifest_size_bytes())
+        .max_blob_size_bytes(global.max_blob_size_bytes())
         .validate_manifest_references(validate_manifest_references)
         .global_immutable_tags(global.immutable_tags)
         .global_immutable_tags_exclusions(global.immutable_tags_exclusions.clone());